+++ src/pairs.ts
